+++ Project.toml
@@ -4,20 +4,18 @@ authors = ["Chris Coey <coey.chris@gmail.com>"]
 version = "0.0.0"
 
 [deps]
-Combinatorics = "861a8166-3701-5b0c-9a16-15d98fcdc6aa"
-DynamicPolynomials = "7c1d4256-1411-5781-91ec-d7bc3513ac07"
-FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
+DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 GLPK = "60bf3e95-4087-53dc-ae20-288a0d20c6a6"
-GSL = "92c85e6c-cbff-5e0c-80f7-495c94daaecd"
-Gurobi = "2e9cd046-0924-5485-92f1-d5272153d98b"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
-LinQuadOptInterface = "f8899e07-120b-5979-ab1d-7b97bb9e1a48"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MathOptInterface = "b8f27783-ece8-5eb3-8dc8-9495eed66fee"
-PolyJuMP = "ddf597a6-d67e-5340-b84c-e37d84115374"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-SemialgebraicSets = "8e049039-38e8-557d-ae3a-bc521ccf6204"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
-SumOfSquares = "4b9e565b-77fc-50a5-a571-1244f986bda1"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[compat]
+DocStringExtensions = "0.8"
+MathOptInterface = "0.10"
+JuMP = "0.22"
+julia = "1.6"
