+++ README.md
@@ -2,4 +2,4 @@
 
 Under construction. Only works on Julia master.
 
-A polyhedral approximation solver for general mixed-integer convex conic optimization problems. An extension of methods in Pajarito.
+A polyhedral approximation solver for general mixed-integer convex conic optimization problems.
