--- src/Solvers/Solvers.jl
@@ -1,28 +1,0 @@
-#=
-Copyright 2019, Chris Coey and contributors
-
-functions and caches for polyhedral approximation algorithms
-=#
-
-module Solvers
-
-using Aspasia
-using Printf
-using LinearAlgebra
-import MathOptInterface
-const MOI = MathOptInterface
-const MOIU = MathOptInterface.Utilities
-
-abstract type Solver end
-
-# separation cuts algorithm
-include("separation_cuts/solver.jl")
-
-get_status(solver::Solver) = solver.status
-get_solve_time(solver::Solver) = solver.solve_time
-get_num_iters(solver::Solver) = solver.num_iters
-
-get_obj_value(solver::Solver) = solver.obj_value
-get_obj_bound(solver::Solver) = solver.obj_bound
-
-end
